+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api.hs
@@ -17,10 +17,13 @@ module Cardano.Ledger.Api (
   module Cardano.Ledger.Api.PParams,
   -- | Scripts
   module Cardano.Ledger.Api.Scripts,
+  -- | Governance
+  module Cardano.Ledger.Api.Governance,
 )
 where
 
 import Cardano.Ledger.Api.Era
+import Cardano.Ledger.Api.Governance
 import Cardano.Ledger.Api.PParams
 import Cardano.Ledger.Api.Scripts
 import Cardano.Ledger.Api.Tx